Allow vyper SC verifiaction and refactoring
